+++ src/components/Modals/BurgerModal/BurgerModal.jsx
@@ -6,11 +6,11 @@ import sprite from "../../../assets/icons/sprite.svg";
 
 import styles from "./BurgerModal.module.css";
 
-export const BurgerModal = ({ onBurgerMenuClick }) => {
+export const BurgerModal = ({ open, onClose }) => {
   useEffect(() => {
     const onEscPress = (evt) => {
       if (evt.code === "Escape") {
-        onBurgerMenuClick();
+        onClose();
       }
     };
 
@@ -19,39 +19,43 @@ export const BurgerModal = ({ onBurgerMenuClick }) => {
     return () => {
       document.removeEventListener("keydown", onEscPress);
     };
-  }, [onBurgerMenuClick]);
+  }, [onClose]);
 
   return (
-    <div className={styles.backdrop}>
-      <div className={styles.header}>
-        <Logo />
-        <button
-          className={styles.closeBtn}
-          type="button"
-          onClick={() => onBurgerMenuClick()}
-        >
-          <svg className={styles.closeSvg}>
-            <use xlinkHref={`${sprite}#close`} />
-          </svg>
-        </button>
-      </div>
-      <ul className={styles.headerNavList}>
-        <li className={styles.headerNavItem}>
-          <NavLink to="/" className={styles.homeLink}>
-            Home
-          </NavLink>
-        </li>
-        <li className={styles.headerNavItem}>
-          <NavLink
-            to="/recipe"
-            className={styles.recipeLink}
-            state={{ recipe: "Add recipe" }}
-          >
-            Add recipe
-          </NavLink>
-        </li>
-      </ul>
-      <HeroImageContainer />
-    </div>
+    <>
+      {open && (
+        <div className={styles.backdrop}>
+          <div className={styles.header}>
+            <Logo />
+            <button
+              className={styles.closeBtn}
+              type="button"
+              onClick={() => onClose()}
+            >
+              <svg className={styles.closeSvg}>
+                <use xlinkHref={`${sprite}#close`} />
+              </svg>
+            </button>
+          </div>
+          <ul className={styles.headerNavList}>
+            <li className={styles.headerNavItem}>
+              <NavLink to="/" className={styles.homeLink}>
+                Home
+              </NavLink>
+            </li>
+            <li className={styles.headerNavItem}>
+              <NavLink
+                to="/recipe"
+                className={styles.recipeLink}
+                state={{ recipe: "Add recipe" }}
+              >
+                Add recipe
+              </NavLink>
+            </li>
+          </ul>
+          <HeroImageContainer />
+        </div>
+      )}
+    </>
   );
 };
