+++ src/components/layout/Header/BurgerMenu/BurgerMenu.jsx
@@ -3,10 +3,14 @@ import { Button } from "../../../shared";
 import sprite from "../../../../assets/icons/sprite.svg";
 import styles from "./BurgerMenu.module.css";
 
-export const BurgerMenu = ({ isHomePage }) => {
+export const BurgerMenu = ({ isHomePage, handleMenuToggle }) => {
   return (
     <div className={styles.burgerMenu}>
-      <Button className={styles.burgerMenuButton} type="button">
+      <Button
+        onClick={() => handleMenuToggle()}
+        className={styles.burgerMenuButton}
+        type="button"
+      >
         <svg
           width="28"
           height="28"
