+++ src/components/layout/Header/Header.jsx
@@ -8,6 +8,7 @@ import { BurgerMenu } from "./BurgerMenu";
 import { AuthButtons } from "./AuthButtons";
 import { UserBar } from "./UserBar";
 import styles from "./Header.module.css";
+import { BurgerModal, LogOutModal } from "../../Modals";
 
 export const Header = ({ isHomePage }) => {
   const [isLoggedIn] = useState(false);
@@ -16,6 +17,16 @@ export const Header = ({ isHomePage }) => {
   const handleMenuToggle = () => {
     setIsMenuOpen(!isMenuOpen);
   };
+
+  // delete me start
+  const [open, setOpen] = useState(false);
+  const onClickModal = () => {
+    setOpen(true);
+  };
+  const onClose = () => {
+    setOpen(false);
+  };
+  // delete me end
 
   return (
     <header
@@ -34,9 +45,20 @@ export const Header = ({ isHomePage }) => {
             </div>
           ) : (
             <>
+              {/* delete me start */}
+              <button type="button" onClick={onClickModal}>
+                modal
+              </button>
+              <LogOutModal open={open} onClose={onClose} />
+              {/* delete me start */}
+
               <HeaderNav />
               <UserBar />
-              <BurgerMenu isHomePage={isHomePage} onClick={handleMenuToggle} />
+              <BurgerMenu
+                isHomePage={isHomePage}
+                handleMenuToggle={handleMenuToggle}
+              />
+              <BurgerModal open={isMenuOpen} onClose={handleMenuToggle} />
             </>
           )}
         </div>
